parsing/requirements.txt
# ---- API stack (Pydantic v1 compatible) ----
fastapi==0.95.2
starlette==0.27.0
pydantic==1.10.13
uvicorn[standard]==0.22.0

# ---- NLP stack ----
spacy==3.7.4
sentence-transformers==3.0.1
rapidfuzz==3.9.1
pymupdf==1.24.9
python-docx==1.1.2
numpy==1.26.4
